+++ mean_field/mex/calcF.cpp
@@ -11,7 +11,7 @@
 #define NUM_AA_FEATS 210
 
 // Get index of mu_ij given seqence length. Have to offset for upper triangular matrix
-inline size_t get_idx(size_t seqlen, size_t i, size_t j) {
+inline int get_idx(int seqlen, int i, int j) {
 	return i*seqlen - (i+1)*(i+2)/2 + j;
 }
 
@@ -26,15 +26,15 @@ inline size_t get_idx(size_t seqlen, size_t i, size_t j) {
  */ 
 double calcF(
 	const double *mus, 
-	const uint32_t *feats_aa, 
-	const size_t seqlen,
+	const int *feats_aa, 
+	const int seqlen,
 	const double *theta_tri, 
 	const double *gamma, 
 	const double theta_dist, 
 	const double theta_seqlen, 
 	const double theta_prior, 
 	double *gradF,
-	uint32_t condition_dist) {
+	int condition_dist) {
 
 	double F = 0;
 	double mu_ij, mu_ik, mu_jk;
@@ -43,8 +43,8 @@ double calcF(
 	double seq_feat = seqlen*theta_seqlen;
 
 	// i, j are less equals, but k is less than to deal with sum over two versus sum over three elements
-	for (size_t i = 0; i <= seqlen-2; i++) {
-		for (size_t j = i+1; j <= seqlen-1; j++) {
+	for (int i = 0; i <= seqlen-2; i++) {
+		for (int j = i+1; j <= seqlen-1; j++) {
 			//conditioning
 			if (j - i > condition_dist){
 				mu_ij = mus[get_idx(seqlen, i, j)];
@@ -59,7 +59,7 @@ double calcF(
 				gradF[NUM_INTERACTIONS + NUM_AA_FEATS + 2] += mu_ij; // prior
 
 				// Calculation for triplet factors. Depends on mu_ij, mu_ik, mu_jk
-				for (size_t k = j+1; k < seqlen; k++) {
+				for (int k = j+1; k < seqlen; k++) {
 					if (k - i > condition_dist){ //if all three are too close, then don't do anything
 						mu_ik = mus[get_idx(seqlen, i, k)];
 						mu_jk = mus[get_idx(seqlen, j, k)];
@@ -92,14 +92,14 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
 		mexErrMsgTxt("calcF: Arguments 2 and 3 must be UINT32.");
 
 	double *mus = mxGetPr(prhs[0]);
-	uint32_t *feats_aa = (uint32_t *) mxGetData(prhs[1]);
-	uint32_t seqlen = *((uint32_t *) mxGetData(prhs[2]));
+	int *feats_aa = (int *) mxGetData(prhs[1]);
+	int seqlen = *((int *) mxGetData(prhs[2]));
 	double *theta_tri = mxGetPr(prhs[3]);
 	double *gamma = mxGetPr(prhs[4]);
 	double theta_dist = *(mxGetPr(prhs[5]));
 	double theta_seqlen = *(mxGetPr(prhs[6]));
 	double theta_prior = *(mxGetPr(prhs[7]));
-	const uint32_t condition_dist = *((uint32_t *) mxGetData(prhs[8]));
+	const int condition_dist = *((int *) mxGetData(prhs[8]));
 
 	// char buf[100];
 	// sprintf(buf, "Condition distance: %u \n", condition_dist);
@@ -108,7 +108,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
 
 
 	double *gradF = (double *) mxCalloc(NUM_FEATURES, sizeof(double));
-	for (size_t i = 0; i < NUM_FEATURES; i++) {
+	for (int i = 0; i < NUM_FEATURES; i++) {
 		gradF[i] = 0;
 	}
